+++ frontend/patient-app/src/pages/ApplyForTherapistPage.tsx
